+++ lib/components/template_outlet/bs_template_outlet.dart
@@ -1,6 +1,6 @@
-import "package:angular2/angular2.dart";
-import 'package:angular2/src/facade/lang.dart';
 import 'dart:async';
+
+import "package:angular2/angular2.dart";
 
 @Directive(selector: "template[bsTemplateOutlet]")
 class BsTemplateOutletDirective {
@@ -15,10 +15,10 @@ class BsTemplateOutletDirective {
   @Input()
   set bsTemplateOutlet(TemplateRef templateRef) {
     new Future(() {
-      if (isPresent(this._insertedViewRef)) {
+      if (this._insertedViewRef != null) {
         _viewContainerRef.remove(this._viewContainerRef.indexOf(this._insertedViewRef));
       }
-      if (isPresent(templateRef)) {
+      if (templateRef != null) {
         _insertedViewRef = _viewContainerRef.createEmbeddedView(templateRef);
         _insertedViewRef.setLocal('\$implicit', ngOutletContext);
       }